# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digital Marketplace API",
        "description": "API documentation for Digital Marketplace - A platform for creators to sell digital products",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/v1/products/{productId}/generate-download-url": {
            "post": {
                "tags": [
                    "Downloads"
                ],
                "summary": "Generate temporary download URL",
                "description": "Generate a temporary signed URL for downloading a purchased product. URL expires in 60 minutes. Customer only.",
                "operationId": "8575f85c1d046bb0652f5829f5b20748",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "expiration_minutes": {
                                        "description": "URL expiration time in minutes (default: 60, max: 1440)",
                                        "type": "integer",
                                        "example": 60
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Download URL generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DownloadUrlResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Product not purchased or Customer role required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You have not purchased this product."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/download/{productId}": {
            "get": {
                "tags": [
                    "Downloads"
                ],
                "summary": "Download product file",
                "description": "Download a purchased product file using a signed URL. URL must be valid and not expired. Customer only.",
                "operationId": "84096497461f0677932c0190cbf8f8ec",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "expires",
                        "in": "query",
                        "description": "Expiration timestamp (auto-generated by signed URL)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1706328990
                        }
                    },
                    {
                        "name": "signature",
                        "in": "query",
                        "description": "URL signature (auto-generated by signed URL)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abc123..."
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File download started",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated or invalid signature",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid signature."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Product not purchased or URL expired",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You have not purchased this product."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product or file not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/products/{productId}/download-info": {
            "get": {
                "tags": [
                    "Downloads"
                ],
                "summary": "Get download information",
                "description": "Get information about a downloadable product including file details and download statistics. Customer only.",
                "operationId": "1714f0401ad5126369b489cf1fdd9d07",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Download information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DownloadInfo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Product not purchased or Customer role required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You have not purchased this product."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Unauthenticated": {
                "title": "Unauthenticated",
                "description": "Unauthenticated response",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "NotFound": {
                "title": "Not Found",
                "description": "Resource not found response",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found"
                    }
                },
                "type": "object"
            },
...
            "DownloadUrlResponse": {
                "title": "Download URL Response",
                "description": "Response containing temporary signed download URL",
                "properties": {
                    "download_url": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/download/5?expires=1706328990&signature=abc123..."
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-27T04:16:30.000000Z"
                    },
                    "expires_in_minutes": {
                        "type": "integer",
                        "example": 60
                    }
                },
                "type": "object"
            },
            "DownloadInfo": {
                "title": "Download Information",
                "description": "Information about a downloadable product",
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "product_title": {
                        "type": "string",
                        "example": "Advanced Laravel Course"
                    },
                    "file_name": {
                        "type": "string",
                        "example": "course-materials.zip"
                    },
                    "file_size": {
                        "type": "integer",
                        "example": 1048576
                    },
                    "file_size_formatted": {
                        "type": "string",
                        "example": "1.0 MB"
                    },
                    "mime_type": {
                        "type": "string",
                        "example": "application/zip"
                    },
                    "purchased_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-27T03:16:30.000000Z"
                    },
                    "download_stats": {
                        "properties": {
                            "total_downloads": {
                                "type": "integer",
                                "example": 3
                            },
                            "first_download": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2026-01-27T03:20:00.000000Z",
                                "nullable": true
                            },
                            "last_download": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2026-01-28T10:15:00.000000Z",
                                "nullable": true
                            },
                            "recent_downloads": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "downloaded_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "ip_address": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}